% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{post_plot}
\alias{post_plot}
\title{Posterior density plot}
\usage{
post_plot(trial)
}
\arguments{
\item{trial}{An object of class 'trial'}
}
\value{
5 snapshots of the posterior density plots for the effect sizes through teh trial
}
\description{
Posterior density plot
}
