% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{psup_plot}
\alias{psup_plot}
\title{Probability of superiority plot}
\usage{
psup_plot(trial)
}
\arguments{
\item{trial}{An object of class 'trial'}
}
\value{
Probabilities of superiority for each arm evolving through the course of the trial
}
\description{
Probability of superiority plot
}
