% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAR_sim.R
\name{RAR_sim}
\alias{RAR_sim}
\title{RAR simulator}
\usage{
RAR_sim(nt, theta0, nb = 1, maxN = 500, N = 1000, upper = 0.95,
  lower = 0.05, burn = 10 * nt, response.type = "binary")
}
\arguments{
\item{nt}{Number of treatment arms}

\item{theta0}{vector of effect sizes}

\item{nb}{batch size for each update}

\item{maxN}{maximum number of patients that the trial can run for; the trial is stopped when \code{maxN} is achieved}

\item{N}{Number of particles}

\item{upper}{upper threshold for probability of superiority; the trial is stopped when one of the trial arms reaches
this value}

\item{lower}{lower threshold for probability of superiority; an arm is stopped when its corresponding probability of
superiority falls below this value}

\item{burn}{the size of pre-adaptation sample; the default is 10 times number of treatments}

\item{response.type}{response type, either 'binary' or 'count'}
}
\value{
updated posterior samples
}
\description{
This function simulates a RAR trial with SMC updates
}
